# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/posts":{"get":{"operationId":"PostsController_findAll","parameters":[],"responses":{"200":{"description":"Posts retrieved successfully."}},"tags":["posts"]},"post":{"operationId":"PostsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostModel"}}}},"responses":{"201":{"description":"Post created successfully."},"422":{"description":"Post title already exists."}},"tags":["posts"]}},"/posts/{id}":{"get":{"operationId":"PostsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Post retrieved successfully."},"404":{"description":"Post not found."}},"tags":["posts"]},"delete":{"operationId":"PostsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":"Post created successfully."},"422":{"description":"Post title already exists."}},"tags":["posts"]},"put":{"operationId":"PostsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostModel"}}}},"responses":{"200":{"description":"Post updated successfully."},"404":{"description":"Post not found."},"422":{"description":"Post title already exists."}},"tags":["posts"]}}},"info":{"title":"Bike Renting API","description":"REST API đơn giản cho việc lưu trữ thông tin thuê xe","version":"1.0","contact":{}},"tags":[{"name":"Bike","description":""}],"servers":[],"components":{"schemas":{"PostModel":{"type":"object","properties":{"id":{"type":"number"},"date":{"type":"string","format":"date-time"},"bikeID":{"type":"number"},"rentingTime":{"type":"number"},"total":{"type":"number"}},"required":["date","bikeID","rentingTime","total"]}}}}